cherry-status/requirements.txt
gunicorn
cherrypy
psycopg2-binary
sqlalchemy-json
sqlalchemy-timescaledb
jinja2
requests
argon2-cffi
